requirements.txt
python-dotenv
streamlit
huggingface_hub
torch
